add hover tooltips on thumbnails
This feature's meant for grid view, but one issue: it kinda bleeds out
to the column and expanded view as well. I'm not sure how to keep it
for that view without doing a lot of DOM wizardry w/ JS (I don't wanna
do that just yet, for performance concerns).