throttle mark-reading on scrolling in-/out of #tiles
The `lastmax` check was meant to be a throttle for updateStorageState()
calls. Because `lastmax` gets reassigned again later on, any additional
sketches that get added that should be marked as "new!" don't get marked
as such. Therefore, the check had to be removed.

Scrolling in and out of #tiles seems to be a good enough throttle, and it
makes a bit more sense to mark the gallery as read when we've just seen the
latest sketch on it.