replace `:first-child` CSS selectors with `.tile-image-holder`
For semantics' sake.

Surprisingly, this also fixes the issue with .new-status having a really wide
left padding, probably because .new-status used to fall under the :first-child
selector. Either this commit fixed it or 13f4f77 did. Regardless, the rules that
used to fix it are now removed as a result.