fix canvas crashes on page idle (chrome mobile)
I can't stress how long I've been trying to catch this bug, let alone
find a fix for it. It took me two weeks of debugging and most of it was
spent on waiting for it to happen again. Then I realized that
`chrome://gpucrash` exists, and I thought about those two weeks a bit.