refresh gallery when new sketches are way too many
Chrome's kinda the reason for this -- like Firefox, it also tends to
cache `getStats.php` responses, despite the `no-store` cache directive
on fetch. Sometimes the cached response may be so old that it's a
thousand or so sketches away from present day.

I don't think it's a good idea to extend it all one-by-one.