fix uncentered alignment on individual settings
I think that the `vertical-align: bottom` part was meant to fix the gaps under
the checkboxes when the rule is taken out. And when the <select>s came along,
they started looking a wee bit off with it. Changing it to
`vertical-align: middle` should quickly fix that up.

The preferences now look a bit more padded apart though...