fix 404 tiles not assigning `first` on extendTo()
This should fix a bug where clicking on the "show more above" button on
a tile list that's full of 404 tiles jumps the page to ID # 10. `first`
is null, and on the button's click listener, `first + 10` would return
10.

Same fix will be applied to extendAbove() with `last`, just in case that
assignment ever starts to matter there.