anchor extendAbove() elements to topmost gallery tile
This fixes a crash with extendAbove() where on a page full of 404 tiles,
there would be no .tile-section present for any new tiles to anchor to.
This commit fixes that by anchoring them to the first .tile instead,
which should always be present as long as extendTo() was fired once.

Sections would be anchored to that .tile too if there's no .tile-section
at the top.